requirements.txt
botbuilder-core>=4.5.0.b4
botbuilder-schema==4.5.0b4
flask>=1.0.3
requests